moved helper function to utils
